refactor(payment): Adjust payment folders and variables colors
