+++ src/app/payment/components/list-payment/table-payment/table-payment.component.scss
