+++ src/styles.scss
@@ -1,5 +1,3 @@
-@use '~@angular/material' as mat;
-
 @import 'src/theme/fonts';
 @import 'src/theme/material';
 @import 'src/theme/variables';
