+++ src/theme/_variables.scss
@@ -1,7 +1,5 @@
-@import 'material';
-
-$primary: mat-color($primary);
-$accent: mat-color($accent);
-$warn: mat-color($warn);
+$primary: #007dfe;
+$accent: #002d69;
+$warn: #f44336;
 $background-color: #f2f2f2;
 $text: #333333;
